GH-4285: move serveGzip into BackendApp method
